Implementation des certain services pour les notifications (objets trouvés)

+++ inquirio-server/src/main/java/ca/obrassard/inquirioCommons/FoundRequest.java
@@ -1,8 +1,8 @@
 package ca.obrassard.inquirioCommons;
 
 public class FoundRequest {
-    public long itemID;
-    public long senderID;
+    public int itemID;
+    public int senderID;
     public String message;
     public byte[] image;
 }
